+++ src/molecules/tanstack-table/utils/column-names2.ts
